Move SequenceReader into our own namespace
It is no longer replaceable by what is included in netcoreapp3.1
