+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/SequenceReader.cs
@@ -5,10 +5,12 @@
  * The .NET Foundation licenses this file to you under the MIT license.
  * See the LICENSE file in the project root for more information. */
 
+using System;
+using System.Buffers;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 
-namespace System.Buffers
+namespace MessagePack
 {
     internal ref partial struct SequenceReader<T>
         where T : unmanaged, IEquatable<T>
